+++ src/views/Flows/Settings/YmlContainer/container.scss
@@ -1,0 +1,7 @@
+@import 'variables/colors';
+.container {
+  margin: 20px 30px;
+  border: 1px solid $color-hr;
+  border-radius: 3px;
+}
+
