+++ src/views/Flows/Settings/YmlContainer/index.js
@@ -1,0 +1,2 @@
+import Container from './container'
+export default Container
